:adhesive_bandage: Remove IntelliJ specific rules from .gitignore
The IntelliJ specific rules in the .gitignore file were removed. This change was made because these rules were unnecessary; they were specific to the developer's local environment and thus didn't need to be kept in the common .gitignore file. By removing them, the .gitignore file becomes cleaner and easier to manage.