:sparkles: Add user application layer to bokeh application
Several new files and dependencies were added to form a user management layer. The LoadUserPort interface is created to load user information whereas CreateUserUseCase and GetUserInfoQuery interfaces are added to handle user data related operations. The aim is to allow the system to manage and interact with user data more effectively. Additional dependencies are added in build.gradle.kts to support the implementation.