:recycle: Refactor user creation and retrieval methods
Deleted the CreateUserUseCase and UseCase classes as part of cleaning up the structure of the codebase and improving legibility. New changes introduce more clear separation of command and query responsibilities.
Created new classes CreateUserParam, UserQueryService, SaveUserPort, and UserCommandService to handle user creation and information retrieval functionality. This ensures that the duties are distributed among more specific methods and classes, and boosts readability.
A default user role has been assigned in the User class as it simplifies user creation process.
Changes have also been made to existing classes where needed, largely relating to modifications in method parameters and return types as per new classes and methods. This is important to ensure consistency in data types across the application. The changes aim to improve the code's maintainability and adaptiveness to future modifications.