:wrench: Update Java version and refactor build.gradle.kts
The 'bokeh-user-server/build.gradle.kts' file was updated to enhance project consistency and maintainability. Specifically, the "java" plugin was replaced with the "java-library" plugin. The Java source compatibility is also updated to version 17, encapsulating the setting inside a 'toolchain' block. This improves readability and provides a more explicit declaration of Java version in this Gradle Kotlin DSL script. The redundant parts (compile tasks and arguments for Kotlin compiler) were also removed as they are not needed anymore.