:sparkles: Add JPA support to Bokeh user service
Updated the `build.gradle.kts` of bokeh-user-server and bokeh-user-framework to implement JPA support. The `spring-boot-starter-data-jpa` dependency was added to provide the necessary JPA classes and the Kotlin JPA plugin was included in the bokeh-user-server to facilitate the generation of no-arg constructors for JPA entities. This integration aims to simplify database operations within the application.