:sparkles: Add User persistence layer
This commit brings major changes to the structure and functionality of the user service. Old porter interfaces `LoadUserPort` and `SaveUserPort` are no longer needed, so they've been removed. To replace them, a new `UserPort` interface has been added to handle operations related to User entities.

The framework is refactored to introduce `UserEntity` and `UserEntityMapper` for data conversion, along with a `UserRepository` for data access. New components are adequately annotated for Spring component detection.

Additional user information is now captured in `UserInfo` with the inclusion of `role`.

`UserPersistenceAdapter` is created to hook user application layer to the new framework layer with mapping and repository utilities.

Lastly, A `NotFoundException` is added for better error handling while retrieving Users from the database.

These changes implement a cleaner, more maintainable architecture, and better separation between the application and framework layer, following a hexagonal design.