:sparkles: Add user presentation layer and corresponding DTOs
New files have been added that define the UserRestController, along with the necessary DTOs(UserInfoResponse, UserCreateRequest) for handling user-related requests. The controller includes methods for creating a new user and retrieving user information. The addition of these files is a part of the task to build the bootstrap for the user-related functionality in the bokeh application.