:sparkles: Add email-based user lookup functionality
This commit introduces the ability to lookup users by their email address, in addition to their UUID. This feature was added due to the requirement of having more ways to locate user information, particularly for situations where the UUID is not readily available or convenient. It involved modifying several classes to support the new lookup method.