Add password match check and login API
A password match check method has been added to the User domain and a new message 'INVALID_PASSWORD' to ErrorMessage constants. Additionally, AuthLoginRequest and AuthLoginResponse DTOs were created with corresponding attributes. New files for UserTokens and for AuthRestController, AuthCommandUseCase, and AuthCommand were also added to the codebase. All these changes are part of implementing the login API. The login process involves the application checking if entered password matches the user's password, then generates tokens if the match is successful. Token generation is a TODO.