:sparkles: Refactor utility location and add auth library
The BCryptPasswordEncoder utility has been moved from common/util to user/util to better align with its usage area. It's used in both UserAuthCommandUseCase and UserCommandUseCase. This commit will improve the organization of the code-base.

Furthermore, an authentication library java-jwt is added to the build.gradle.kts for upcoming JWT authentication feature development.