:recycle: Refactor UserAuth token creation and storage methods
Updated the token generation method within UserTokenProvider to return a UserAuthToken object including the userId. This allows for simpler interaction when storing the token in Redis - we can now get the userId directly from the UserAuthToken object, reducing the risk of ID mismatches. The UserAuthToken object is also now for better clarity.