:memo: Add system design and code architecture description to README
A detailed description of the system design and code architecture was included in the README file. This addition is intended to provide an understanding of the core components (such as Domain Hexagon, Application Hexagon, Framework Hexagon, and Bootstrap Hexagon) that form the structure of the system and their respective roles. A reference to the source from where this architecture was adapted is also mentioned.