:wrench: Add .gitignore file at user-server module
Created a .gitignore file to ignore files generated by IntelliJ, MacOS, Gradle, and Kotlin. This is to prevent tracking of unnecessary files like user-specific configurations, generated files, build files, and local cache, which are not relevant to the project source code. It keeps the repository clean and only holds files necessary for project setup and code execution.