:wrench: Add local environment configurations
Added configurations for local environment in application.yaml and docker-compose.yaml. The spring profile is set to 'local' and docker-compose file path has been specified. Also, added a docker-compose file for the 'mongo' service to run in local environment. These changes will allow the server to run in a consistent local environment for development and testing purposes.