:wrench: Add Dockerfile for Bootstrap API
A Dockerfile was added to the bokeh-post-server sub project to build and run the 'bokeh-post-bootstrap-api' module. The Dockerfile uses Gradle Wrapper to build the jar artifact and OpenJDK 17 to run the application. This simplifies the process of deploying and running the application in any environment, as it packages application with all its dependencies.