:green_heart: Add mask-password and registry-type options in ECR Login
This commit adds two options in the GitHub Actions workflow for logging into Amazon ECR. The 'mask-password' option is set to 'true' to hide the Amazon ECR password in the GitHub Actions log, improving security. The 'registry-type' is set as 'public', specifying the type of registry to which we are trying to log in.