:construction_worker: Add User Service CI workflow
A User Service Build and Push Docker Images workflow was added, being executed on the push of main branch and any changes to the workflow or in the 'user-service' directory. The workflow setup includes AWS credentials configuration, fetching ECR url from AWS SSM, login to Amazon ECR and building and pushing the user-service docker image.