:construction_worker: Add caching to GitHub Actions workflow
This commit adds cache to the user-service CI workflow. The main aim of this change is to speed up build times by caching Gradle dependencies. In addition, the `--build-cache` option is added to the `gradlew test` command which helps store outputs so that Gradle doesn’t have to re-execute tasks when inputs haven’t changed.