:green_heart: Update name of GitHub workflow for CI
The name of the GitHub workflow in the post_service_ci.yaml file has been changed from "Post Service Build and Push Docker Images" to "Post Service CI". The change was made to better reflect the purpose of the workflow which is more aimed at continuous integration (CI) not just building and pushing Docker images.