:recycle: Refactor UserTokenProvider to simplify token creation
The UserTokenProvider class was refactored to simplify the creation of access and refresh tokens. The changes involved creating a new function, createUserAuthToken(). This function adds an extra level of abstraction, reducing code repetition and making the code more readable and manageable. The calls to generateToken() were moved into this new function, and the creation of the UserAuthToken was streamlined. This should not affect any existing functionality but simplifies the code and makes it easier to maintain.