:recycle: Refactor UserTokens and UserAuthToken to UserLoginToken and UserToken
The names UserTokens and UserAuthToken have been renamed to UserLoginToken and UserToken respectively, in relevant files across the user-service. This refactoring aims to better represent their actual use in the codebase. They are both mainly used during the process of logging in, so the more specific names will make the code easier to understand for all developers.