:recycle: Refactor UserToken to be a modular and specific
This commit represents a series of changes made to refactor and restructure the UserToken in the user-service of our application. The main goal was to make the UserToken more modular, specific, and easier to handle in other parts of the service.

The UserToken was initially a single class, but it has been split into UserAccessToken and UserRefreshToken. Now, the tokens have their own specific classes, which makes them more identifiable and manageable. These changes can also facilitate better allocation and management of tokens in the Redis repository.

In UserAuthPort interface, the saveRefreshToken method also has been refactored, instead of passing userId and refreshToken separately, now only refreshToken is being passed which carries all the necessary information.

Overall, these changes should help to improve code quality and robustness. These changes are backwards-compatible and shouldn't affect the functionality of the service.