:fire: Remove automatic tag generation from Post model
In this commit, we remove the createWithGeneratedTagsByContent method from the Post model's companion object. This decision was made because auto tag generation based on post content was determined to be outside of the Post domain's responsibility, which should focus only on representing a post's state. It is more appropriate to delegate this task to another module or service, thereby ensuring good separation of concerns.