:sparkles: Add Post handling interfaces and use-case implementation
In this commit, we have created several new files that create interfaces for post creation and handling. These include PostPort.kt and PostCommand.kt, which act as boundaries for outport and inport directions of the Post domain. Furthermore, CreatePostParam.kt is introduced as a data handling class for obtaining and processing post creation parameters. Finally, we've implemented PostCommandUseCase.kt as a use-case for these interactions. This organization streamlines the construction and compartmentalization of application logic for posts, enhancing clarity and maintainability.