:green_heart: Rename and modify user api ci workflow
The user api ci workflow file was renamed to user-api-rest-ci.yaml and its respective occurrences inside the file were modified. This change was necessary for clarity and better representation of the api's nature which is RESTful. Steps ids and aws parameters related to this ci workflow were updated accordingly to maintain consistency. Also, docker commands were updated to use the new id, ensuring successful build and push to ECR.