:recycle: Update application configurations
This update is intended to streamline how configurations are managed within the project. The AppConfig, PersistenceConfig, and RedisConfig classes have been updated with a ComponentScan annotation, which will auto-load beans from specified packages. This change helps to enhance modularity and separation of components. Furthermore, the JpaConfig was renamed to PersistenceConfig for better representation of its role. Lastly, the UserApiApplication was updated to import these configuration classes.