:wrench: Remove application-local.yaml, improve application.yaml
The file application-local.yaml is removed to streamline the codebase, because the configurations it contained are now included in application.yaml. The settings from the deleted file are moved to the new file application.yaml, which are activated depending on their profile. This makes configuration management more cohesive because related settings are defined in the same file.