:wrench: Update Gradle dependencies and configurations
This commit upgrades the SLF4j API dependency to version 2.0.5 and updates existing plugins in the build.gradle.kts. The changes also involve adjusting all project configurations, making updates to align better with Spring Boot and Kotlin task versions. This was done to take advantage of the new features and improvements these newer versions offer. Additionally, the structure of the build files has been slightly reorganized for readability and to align with Kotlin DSL syntax requirements.