:sparkles: Add logout API
A new data class `UserLogoutRequest` was added to the dto package in UserRequest.kt file. To handle this request, a logout method with `@PostMapping` and `@RequestBody` annotations was added in UserRestController.kt. The method currently does not have any implementation (mentioned as TODO). This change is made to provide logout facility to the user which was missing earlier.