:wrench: Update Gradle build file for Kotlin and Spring Boot updates
The Kotlin and Spring Boot plugins were modified in the Gradle build file. The update was necessary to keep compatibility with the latest versions of Kotlin and Spring Boot. The new build includes Kotlin "jvm", "plugin.jpa" and "plugin.spring" plugins, and it has updated version for "org.springframework.boot" and "io.spring.dependency-management". Also added new dependencies for better support of kotlin-based coroutines and tests. This change should enhance the codebase's maintainability and compatibility.