:sparkles: Add factory method to create Tags objects
Created a factory method in Tags.kt allowing easy conversion from a list of strings to a Tags object. This was added in the PostCreateCommand.kt file to convert a list of tag strings to a Tags object when creating a new Post. This simplifies creation and handling of Tags, ensuring only valid ones can be added to a Post.