dReplace MongoDB with Kafka in Docker Compose
This modification is made to replace the existing MongoDB service in the docker-compose.yml of the post-api-rest with the Confluent Platform services Zookeeper and Kafka. The environment variables for the replaced MongoDB are deleted, and new environment variables for Zookeeper and Kafka are defined. Additionally, the MongoDB image is replaced with the latest images for Zookeeper and Kafka. This change is due to Kafka providing higher throughput, better fault-tolerance, and less latency than MongoDB in handling real-time data.