:sparkles: Add KafkaConfig and refactor PostApiRestApplication
This commit introduces a new class `KafkaConfig` in the `common.config` package to handle Kafka related configurations.
Along with this, `PostApiRestApplication` was moved to `api.rest` package for better structuring, and it now imports the newly added `KafkaConfig` class. This restructure ensures proper isolation of API level configurations and adapts Kafka integration for post micro-service.