:recycle: Refactor code for packaging organization and configuration consistency
In this commit, several crucial changes were made to improve the packaging structure and configuration consistencies in the codebase.
Firstly, the UserApiApplication.kt file was renamed and moved to a new package for better organization and readability.
Secondly, the @Repository annotation in UserJpaRepository.kt was removed as it is, @EntityScan and @EnableJpaRepositories were explicitly defined in PersistenceConfig.kt and similar changes are made in  RedisConfig.kt and AppConfig.k