:wrench: Update Kafka dependency scope in build file
Kafka dependency scope in the post adapter module is changed from 'implementation' to 'api'. This change exposes the dependency transitively. It means any module that depends on the post-adapter module will now have access to spring-kafka classes without explicitly declaring the spring-kafka in their build script. This was necessary to improve resource sharing and help in reducing redundant code across the modules.