:sparkles: Implement post creation event use case
The function `create` in `PostCreateCommand.kt` and `PostEventPort.kt` has been renamed to `sendCreateEvent` to more accurately reflect its purpose. This makes the code more readable and descriptive.

Additionally, `PostKafkaAdapter.kt` has been updated to use a KafkaTemplate for sending the post creation event. A new constant `POST_CREATE` has been defined in `KafkaTopic.kt` to provide the topic name.

This change is part of the necessary steps to integrate Kafka into the application and to achieve a more event-driven architecture.