:sparkles: Add Post API(Websocket) application
The commit introduces a new WebSocket application for the Post API. This was done to provide real-time update capabilities, as the new application includes Kafka implementation replacing the previous Mongo adapter. Files created includes a new kotlin file for the application, updated build.gradle.kts for kafka implementation, and application.yaml for the new functional setup. Also, Docker-compose.yaml was added to handle Kafka and zookeeper services. This significantly improves the real-time capability of the post API.