:sparkles: Add Refactor Kafka to use Reactive implementation
Changed the implementation of Kafka in the project from using KafkaTemplate to the ReactiveKafkaProducerTemplate. The KafkaProducerConfig class was created to hold our custom configuration of producer properties. Simultaneously, the kafka gradle dependencies were updated to their latest versions and added the reactor-kafka dependency, which is required for the reactive approach. This overhaul was decided upon to take advantage of the non-blocking, asynchronous nature of the reactive streams API, that would allow our app to handle more requests with fewer threads and less memory, improving performance.