:adhesive_bandage: Set UUID as not null for User and Post entities
The code has been changed to ensure that the UUID for both the User and Post entities are always set at the time of their instantiation. This change is made to eliminate potential NullPointerExceptions that may arise if an instance is created without a valid UUID. Notably, related test cases have been modified to align with this change. By making UUID compulsory for the User and Post entities, we can improve the safety and consistency of our codebase.