:wrench: Add WebSocket adapter to post-api
In this commit, the WebSocket adapter has been added to the post-api project. Please note, the `spring-websocket dependency` was removed from post-api/websocket and instead applied in the new post-adapter-websocket module. It now allows better module encapsulation for web socket implementation which was previously embedded in the post-api directly. The configuration is also updated to include `websocket`. The adapter reference was