:sparkles: Add WebSocket support for post events
This commit introduces support for WebSocket programming to enable real-time, bidirectional communication between the server and the client. WebSocket configuration is set up to facilitate this connection.
A new interface PostSocketPort was implemented for dispatching post creation events.
Additionally, there are new classes PostCreatedEventSendCommand and PostCreatedEventSendCommandUseCase introduced for sending post creation events to the WebSocket.
A Post to the topic.
The WebSocket configuration is imported into the main application.
Plus, a new controller, PostEventController, is added for processing post events pushed from the client.