:sparkles: Add new worker module in post-api and its dependencies
New post-worker module is added in the post-api directory. Necessary project dependencies are included in the build.gradle.kts file of the post-worker module. This new module helps in sharing common codebase within the post-api and aids in better project organization and scalability. The settings.gradle.kts file has been updated to include this module.