:heavy_plus_sign: Add Jackson and Springdoc dependencies, enable Swagger UI
Added Jackson modules for Kotlin, JDK8, and JSR310 to enhance data processing and add type-safety. Added the Springdoc OpenAPI UI starter to provide an interactive API documentation using Swagger UI. This is to make the API more understandable and easier to use for development teams. Swagger UI is enabled for all profiles but disabled for the 'dev' profile.